feat: template editor and delete forms

+++ src/lib/components/property/properties-table.svelte
@@ -7,6 +7,7 @@
   import { Button } from "$lib/components/ui/button";
   import * as DropdownMenu from "$lib/components/ui/dropdown-menu";
   import * as Select from "$lib/components/ui/select";
+  import * as Tooltip from "$lib/components/ui/tooltip";
   import { state as GLOBAL_STATE, State } from "$lib/model";
   import {
     ASelectProperty,
@@ -21,18 +22,22 @@
   import { ChevronDownIcon, PlusIcon, TagIcon, TagsIcon, TextIcon } from "lucide-svelte";
   import { createRender, FlatColumn, type ReadOrWritable } from "svelte-headless-table";
   import { createSortKeysStore, type WritableSortKeys } from "svelte-headless-table/plugins";
-  import { type Writable, writable } from "svelte/store";
+  import { type Writable, writable, get as _get } from "svelte/store";
   import { v4 as uuidv4 } from "uuid";
   import Label from "../ui/label/label.svelte";
   import Separator from "../ui/separator/separator.svelte";
   import OptionsTable from "./options-table.svelte";
   import PropertyOptions from "./property-options.svelte";
   import { describeType } from "./misc";
+  import DeleteDialog from "../data-table/lib/delete-dialog.svelte";
 
   export let data: ReadOrWritable<Property<unknown>[]>;
   export let state: State = GLOBAL_STATE;
   export let header = true;
   export let rowActions = new Map<string, (item: Property<unknown>) => void>();
+  export let extraColumns: ColumnInitializer<Property<unknown>>[] = [];
+  export let hiddenColumns: string[] = [];
+  export let onEditSubmit: (item: Property<unknown>) => void = () => {};
 
   let filterValue: Writable<string> = writable("");
   let sortKeys: WritableSortKeys = createSortKeysStore([]);
@@ -42,6 +47,7 @@
 
   let nextType: PropertyType = "multiple";
   let selected: Property<unknown> | undefined = undefined;
+  let alertOpen = false;
   let dialogOpen = false;
   let dialogTitle = "Edit Option";
   let oldProps: Partial<SelectPropertyProps> = {};
@@ -81,9 +87,11 @@
       header: "Description",
       id: "description",
     },
+    ...extraColumns,
   ];
 
-  let actions = new Map([...rowActions, ["Edit", rowClick], ["Delete", rowDelete]]);
+  hiddenColumns.forEach((id) => (hideForId[id] = true));
+  let actions = new Map([["Edit", rowClick], ...rowActions, ["Delete", rowDelete]]);
 
   function rowClick(item: Property<unknown>) {
     dialogTitle = "Edit Property";
@@ -92,7 +100,12 @@
   }
 
   function rowDelete(item: Property<unknown>) {
-    item.delete();
+    if (_get(state.settings).askDeleteConfirmation) {
+      selected = item;
+      alertOpen = true;
+    } else {
+      item.delete();
+    }
   }
 
   function newProperty(type: PropertyType, props?: Partial<SelectPropertyProps>, uuid?: string) {
@@ -142,10 +155,6 @@
     dialogOpen = true;
   }
 
-  function onOpenChange() {
-    oldProps = {};
-  }
-
   function changePropType(v: unknown) {
     if (!v) return;
     console.log("Old Props", oldProps);
@@ -167,6 +176,16 @@
     dialogTitle = "Edit Property";
   }
 
+  function onOpenChange() {
+    oldProps = {};
+  }
+
+  function onSubmit(prop: Property<unknown>) {
+    onEditSubmit(prop);
+    prop.push();
+    oldProps = {};
+  }
+
   export { className as class };
 </script>
 
@@ -179,43 +198,52 @@
           slot="start"
         >
           <slot name="add">
-            <Button
-              variant="ghost"
-              size="icon-xl"
-              class="text-muted-foreground hover:text-primary"
-              on:click={() => newProperty(nextType)}
-            >
-              <PlusIcon />
-            </Button>
+            <slot name="add-default">
+              <Tooltip.Root openDelay={100} closeDelay={100} group="template-editor-add-property">
+                <Tooltip.Trigger>
+                  <Button
+                    variant="ghost"
+                    size="icon-xl"
+                    class="text-muted-foreground hover:text-primary"
+                    on:click={() => newProperty(nextType)}
+                  >
+                    <PlusIcon />
+                  </Button>
+                </Tooltip.Trigger>
+                <Tooltip.Content>Create a new property</Tooltip.Content>
+              </Tooltip.Root>
+            </slot>
             <Separator
               orientation="vertical"
               class="h-7 w-[1px] bg-muted-foreground bg-opacity-50 transition-all group-hover:bg-transparent"
             />
-            <DropdownMenu.Root>
-              <DropdownMenu.Trigger
-                class="flex h-11 w-11 items-center justify-center rounded-md text-muted-foreground transition-all hover:bg-accent hover:text-primary"
-              >
-                <ChevronDownIcon />
-              </DropdownMenu.Trigger>
-              <DropdownMenu.Content>
-                <DropdownMenu.Group>
-                  <DropdownMenu.Label class="text-center">Choose a Type</DropdownMenu.Label>
-                  <DropdownMenu.Separator />
-                  <DropdownMenu.Item on:click={() => newProperty("single")}>
-                    <TagIcon class="mr-2 w-8 text-muted-foreground" />
-                    {describeType("single")}
-                  </DropdownMenu.Item>
-                  <DropdownMenu.Item on:click={() => newProperty("multiple")}>
-                    <TagsIcon class="mr-2 w-8 text-muted-foreground" />
-                    {describeType("multiple")}
-                  </DropdownMenu.Item>
-                  <DropdownMenu.Item on:click={() => newProperty("text")}>
-                    <TextIcon class="mr-2 w-8 text-muted-foreground" />
-                    {describeType("text")}
-                  </DropdownMenu.Item>
-                </DropdownMenu.Group>
-              </DropdownMenu.Content>
-            </DropdownMenu.Root>
+            <slot name="add-dropdown">
+              <DropdownMenu.Root>
+                <DropdownMenu.Trigger
+                  class="flex h-11 w-11 items-center justify-center rounded-md text-muted-foreground transition-all hover:bg-accent hover:text-primary"
+                >
+                  <ChevronDownIcon />
+                </DropdownMenu.Trigger>
+                <DropdownMenu.Content>
+                  <DropdownMenu.Group>
+                    <DropdownMenu.Label class="text-center">Choose a Type</DropdownMenu.Label>
+                    <DropdownMenu.Separator />
+                    <DropdownMenu.Item on:click={() => newProperty("single")}>
+                      <TagIcon class="mr-2 w-8 text-muted-foreground" />
+                      {describeType("single")}
+                    </DropdownMenu.Item>
+                    <DropdownMenu.Item on:click={() => newProperty("multiple")}>
+                      <TagsIcon class="mr-2 w-8 text-muted-foreground" />
+                      {describeType("multiple")}
+                    </DropdownMenu.Item>
+                    <DropdownMenu.Item on:click={() => newProperty("text")}>
+                      <TextIcon class="mr-2 w-8 text-muted-foreground" />
+                      {describeType("text")}
+                    </DropdownMenu.Item>
+                  </DropdownMenu.Group>
+                </DropdownMenu.Content>
+              </DropdownMenu.Root>
+            </slot>
           </slot>
           <slot name="start" />
         </div>
@@ -243,13 +271,8 @@
     />
   </div>
 </div>
-<EditDialog
-  item={selected}
-  bind:open={dialogOpen}
-  title={dialogTitle}
-  onSubmit={(opt) => opt.push()}
-  {onOpenChange}
->
+
+<EditDialog {selected} bind:open={dialogOpen} title={dialogTitle} {onSubmit} {onOpenChange}>
   <svelte:fragment slot="options">
     {#if selected}
       <div class="flex w-full flex-col gap-1.5">
@@ -274,6 +297,13 @@
           <OptionsTable property={selected} />
         </div>
       {/if}
+      <slot name="editForm" />
     {/if}
   </svelte:fragment>
 </EditDialog>
+
+<DeleteDialog
+  {selected}
+  open={alertOpen}
+  extraDescription="This will remove all references to this property in all items."
+/>
